requirements.txt
Flask>=3.0,<4.0
mysql-connector-python>=8.4,<9.0
gunicorn>=22.0,<23.0
pytest>=8.0,<9.0
